+++ lib/router.dart
@@ -7,7 +7,8 @@ import 'package:package_info_plus/package_info_plus.dart';
 import 'screens/update_required.dart';
 import 'package:flutter/foundation.dart';
 import 'package:http/http.dart' as http;
-import 'services/user_data_manager.dart' show backendBaseUrl, updateUtcOffset;
+import 'services/user_data_manager.dart'
+    show backendBaseUrl, updateUtcOffset, authenticatedPost;
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:go_router/go_router.dart';
 import 'home_screen.dart';
@@ -527,6 +528,7 @@ class _AppInitScreenState extends ConsumerState<AppInitScreen> {
     await workoutNotifier.checkAndRestoreWorkoutSession();
 
     updateUtcOffset();
+    if (!isGuest) authenticatedPost('app_open', body: {});
 
     final userData = notifier.currentUserData;